agents/market_researcher.py
"""Market Researcher performs market research, extracts relevant text snippet and stores results"""

from crewai import Agent, Task
from crewai.tools import tool
import yfinance as yf
from config.settings import get_config
from tavily import TavilyClient
from langchain_openai import ChatOpenAI
from .base import load_prompt

@tool("Get Market Data")
def get_market_data(ticker: str) -> str:
    """
    Fetches the current market data for a given ticker symbol.
    
    Use this tool when you need to find the current market data.
    Input should be a valid stock ticker symbol (e.g., 'AAPL' for Apple).
    
    Args:
        ticker: A stock ticker symbol (uppercase recommended)
    
    Returns:
        A string with the current market data or an error message
    """
    try:
        config = get_config()

        ticker = ticker.strip().upper()
        tavily_client = TavilyClient(config['tavily_api_key'])
        market_query = f"Bring up some of the latest market data for stock {ticker}"

        search_response = tavily_client.search(market_query)
        print(search_response)
        
        market_data = ""
        for result in search_response["results"]:
            market_data += f"### {result['title']}\n\n{result['content']}\n\n"
       
        print("=" * 70)
        print(market_data)
        print("=" * 70 + "\n")

        return f"""
            The current market data for ({ticker})\n
            {market_data}
            """
        
    except Exception as e:
        # GRACEFUL DEGRADATION: Return useful error info instead of crashing
        return f"Error fetching market data for '{ticker}': {str(e)}. Please check the ticker symbol."


def build_market_researcher() -> Agent:
    """
    Build the market researcher agent.

    This agent performs market research, extracts relevant text snippet and stores results in vector memory. 

    Returns:
        Configured market researcher agent
    """
    
    # Load system prompt
    prompt = load_prompt('market_researcher.md')

    agent = Agent(
        role="Market Researcher",
        goal="performs market research, extracts relevant text snippet and stores results",
        backstory=(prompt),
        verbose=True,
        tools=[get_market_data],

        # ALLOW DELEGATION: Set to False because we want this agent to do the
        # research itself, not delegate to the writer (who has no tools anyway).
        allow_delegation=False,
        
       # LLM: We use temperature=0.7 for more creative, engaging writing
        # while still maintaining coherence and accuracy.
        llm=ChatOpenAI(model_name="gpt-4o-mini", temperature=0.7)
    )

    return agent

def build_market_researcher_task() -> Task:
    """
    Build market_researcher task:

    1 Searches markets, news, press releases, and analyst commentary
    2 Extracts relevant text snippets
    3 Stores results in vector memory
   
    Returns:
        Configured market research task
    """
    task = Task(
       description=(
        "You are market researcher who performs the tasks: \n"
        " 1 Searches markets, news, press releases, and analyst commentary\n "
        " 2 Extracts relevant text snippets\n"
        " 3 Stores results in vector memory"
    ),
    expected_output=(
        "A  market research report in professional format with:\n"
        "- A compelling headline\n"
        "- Financial market news, press releases and analyst commentary \n"
        "- Brief analysis\n"
        "- Closing thought"
    ),
    agent=build_market_researcher()    
    )

    return task