requirements.txt
# Core Agent Framework
#crewai==0.80.0 # version 0.80.0 not exist
crewai>=1.6.1
langchain_openai

# Tools & APIs
yfinance              # For stock data
#duckduckgo-search     # Free search tool (good for demos vs. paid Tavily)
tavily-python
python-dotenv         # For managing API keys

# Utilities
ipykernel             # If you want to run this in notebooks later